creating and seperating each choice menu into its respective file
